requirements.txt
pytest==8.2.2
pytz==2024.1
requests==2.32.3
urllib3==2.2.2
google-generativeai==0.7.2
gcloud==0.18.3
oauth2client~=4.1.3
google-api-python-client==2.140.0
google-cloud-storage==2.18.2
google-cloud-secret-manager==2.20.2
functions-framework==3.8.1